# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sport Api Swagger Documentation Anatation",
        "description": "Sport Api Documentation to build the frontend of this site",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "abduvaliyevasad23@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Sport api server link"
        }
    ],
    "paths": {
...
        "/api/pasport": {
            "get": {
                "tags": [
                    "Pasport"
                ],
                "summary": "Get all data from Pasport database",
                "description": "Via this link All Pasport` datas come",
                "operationId": "pasport",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pasport"
                ],
                "summary": "Register a pasport",
                "operationId": "storePasport",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "pnfl": {
                                        "description": "Type your pnfl",
                                        "type": "integer"
                                    },
                                    "pasport_seria": {
                                        "description": "Type your pasport seria",
                                        "type": "string"
                                    },
                                    "pasport_seria_code": {
                                        "description": "Type your pasport seria code",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/pasport/{pasport}": {
            "get": {
                "tags": [
                    "Pasport"
                ],
                "summary": "Get one data from Pasport database",
                "description": "Via this link a Pasport`s data comes to show",
                "operationId": "showPasport",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pasport",
                        "in": "path",
                        "description": "ID for pasport data",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/pasport/{pasport}/update": {
            "put": {
                "tags": [
                    "Pasport"
                ],
                "summary": "Updated",
                "description": "Update this pasport",
                "operationId": "updatePasport",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pasport",
                        "in": "path",
                        "description": "pasport that to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pnfl": {
                                        "description": "Type your pnfl",
                                        "type": "integer"
                                    },
                                    "pasport_seria": {
                                        "description": "Type your pasport seria",
                                        "type": "string"
                                    },
                                    "pasport_seria_code": {
                                        "description": "Type your pasport seria code",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid pasport supplied"
                    },
                    "404": {
                        "description": "pasport not found"
                    },
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/pasport/{pasport}/delete": {
            "delete": {
                "tags": [
                    "Pasport"
                ],
                "summary": "Deletes a pasport",
                "operationId": "destroyPasport",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pasport",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid course supplied"
                    },
                    "404": {
                        "description": "course not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}